+++ VirtualShelter/src/main/java/hu/bme/aut/vshelter/entity/Handicap.java
@@ -15,16 +15,16 @@ public class Handicap {
 	@Id
 	@GeneratedValue
 	@Column(name = "id", unique = true, nullable = false)
-	private int id;
+	private long id;
 	
 	@Column(name = "handicapName", unique = true, nullable = false)
 	private String handicapName;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
@@ -40,7 +40,7 @@ public class Handicap {
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = prime * result + id;
+		result = prime * result + (int)id;
 		return result;
 	}
