+++ VirtualShelter/src/main/java/hu/bme/aut/vshelter/entity/User.java
@@ -49,7 +49,7 @@ public class User extends Advertiser {
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = prime * result + id;
+		result = prime * result + (int)id;
 		return result;
 	}
